requirements.txt
flake8==7.1.1
pytest==8.3.3
requests==2.32.3
pydantic==2.9.2
selenium==4.25.0
allure-pytest==2.13.5